+++ lib/mollie_ex/resources/payment_routes/list.ex
@@ -1,0 +1,51 @@
+defmodule MollieEx.Resources.PaymentRoutes.List do
+  @moduledoc false
+
+  alias MollieEx.Client
+  alias MollieEx.Error
+  alias MollieEx.HTTP.Request
+  alias MollieEx.Resources.PaymentRoutes.Options
+
+  @allowed_options [
+    :testmode,
+    :pool_timeout,
+    :receive_timeout,
+    :request_timeout
+  ]
+
+  @spec build(Client.t(), String.t(), keyword()) ::
+          {:ok, Request.t(), keyword()} | {:error, Error.t()}
+  def build(%Client{} = client, payment_id, opts) when is_binary(payment_id) and is_list(opts) do
+    with :ok <- Options.ensure_keyword(opts),
+         :ok <- Options.reject_unknown(opts, @allowed_options),
+         {:ok, payment_id} <- Options.payment_id(payment_id),
+         {:ok, testmode} <- Options.effective_testmode(client, opts) do
+      request = %Request{
+        method: :get,
+        path: "/payments/" <> Options.encode_path_segment(payment_id) <> "/routes",
+        path_template: "/payments/{paymentId}/routes",
+        query: query(testmode),
+        idempotency_policy: :unsupported,
+        operation: :payment_routes_list,
+        testmode: testmode
+      }
+
+      {:ok, request, Options.timeout_options(opts)}
+    end
+  end
+
+  def build(%Client{}, _payment_id, opts) when not is_list(opts),
+    do: configuration_error(:invalid_options)
+
+  def build(%Client{}, _payment_id, _opts), do: configuration_error(:invalid_payment_id)
+
+  defp query(testmode) do
+    []
+    |> maybe_put(:testmode, testmode)
+  end
+
+  defp maybe_put(query, _key, nil), do: query
+  defp maybe_put(query, key, value), do: Keyword.put(query, key, value)
+
+  defp configuration_error(reason), do: Options.configuration_error(reason)
+end